bootstrap: minimal impl, docs, packaging
As specified in https://github.com/elastic/logstash/issues/15163 ,
this bootstrap creates a new `logstash` integration plugin containing
a `logstash` input plugin that is capable of receiving events from the
also-included `logstash` output plugin.

The plugins are defined and specified in such a way that the protocol
and encoding are both internal implementation details, which allows users
to rely on a simple and consistent interface in their pipelines and for
developers to be free to introduce new and better ways of transmitting
those events in the future.